combine callout and upsell codemods to single function
