--- packages/gestalt-codemods/16.0.0/__testfixtures__/callout-primary-secondary-action-replace-primary-secondary-link.output.js
@@ -1,16 +1,0 @@
-// @flow strict
-import React from 'react';
-import { Callout } from 'gestalt';
-
-export default function TestBox() {
-  return (
-    <Callout
-      message="Insert a clever info callout message here"
-      iconAccessibilityLabel="info"
-      primaryAction={{ href: 'pinterest.com', label: 'Visit Pinterest' }}
-      secondaryAction={{ href: 'pinterest.com/help', label: 'Learn more' }}
-      type="info"
-      title="A Title"
-    />
-  );
-}
