+++ packages/gestalt-codemods/16.0.0/__testfixtures__/callout-upsell-primary-secondary-action-replace-primary-secondary-link.output.js
@@ -1,0 +1,24 @@
+// @flow strict
+import React from 'react';
+import { Box, Callout, Upsell } from 'gestalt';
+
+export default function TestBox() {
+  return (
+    <Box>
+      <Callout
+        message="Insert a clever info callout message here"
+        iconAccessibilityLabel="info"
+        primaryAction={{ href: 'pinterest.com', label: 'Visit Pinterest' }}
+        secondaryAction={{ href: 'pinterest.com/help', label: 'Learn more' }}
+        type="info"
+        title="A Title"
+      />
+      <Upsell
+        message="Insert a clever upsell message here"
+        primaryAction={{ href: 'pinterest.com', label: 'Visit Pinterest' }}
+        secondaryAction={{ href: 'pinterest.com/help', label: 'Learn more' }}
+        title="A Title"
+      />
+    </Box>
+  );
+}
